537: move browse target resolver to dedicated file
